fix double to int conversion warning
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_8_BRANCH@56135 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/dcbase.cpp
@@ -1169,6 +1169,6 @@ float wxDCBase::GetFontPointSizeAdjustment(float dpi)
     // ports need to emulate this bug too:
     const wxSize screenPixels = wxGetDisplaySize();
     const wxSize screenMM = wxGetDisplaySizeMM();
-    const int screenPPI_y = (screenPixels.y * 25.4) / screenMM.y;
-    return float(screenPPI_y) / dpi;
+    const float screenPPI_y = (screenPixels.y * 25.4) / screenMM.y;
+    return screenPPI_y / dpi;
 }
